requirements.txt
PySide6
pillow
pytesseract
requests